Fix bug with an extra space appearing after </a> tags.
This was being caused by the pretty-printing config used by
Text.XML.Light. Pretty-printing added a newline after ending tags, which
for </a> and other closing tags, appeared in the browser as an
additional space after the closing tag because whitespace is collapsed
in the browser.

Fixed by switching from ppContent -> showContent.